API Base - Python/requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
pydantic
# ORM (SQLAlchemy)
sqlalchemy
# Celery e Redis para agendamento
celery
redis
celery[redis]
# Canais de comunicação
twilio
# Utilitários
python-multipart
python-dotenv
# Autenticação
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
python-multipart
email-validator
